+++ libmonit/src/io/OutputStream.h
@@ -120,7 +120,7 @@ bool OutputStream_isClosed(T S);
  * @param S An OutputStream object
  * @return The total number of bytes written
  */
-long long OutputStream_getBytesWritten(T S);
+int64_t OutputStream_getBytesWritten(T S);
 
 //@}
