+++ src/net/Link.h
@@ -91,7 +91,7 @@ void Link_update(T L);
  * @param L A Link object
  * @return Incoming bytes per second or -1 if not available.
  */
-long long Link_getBytesInPerSecond(T L);
+int64_t Link_getBytesInPerSecond(T L);
 
 
 /**
@@ -101,7 +101,7 @@ long long Link_getBytesInPerSecond(T L);
  * range given by 'now - count' (count max = 60m)
  * @return Incoming bytes per minute or -1 if not available.
  */
-long long Link_getBytesInPerMinute(T L, int count);
+int64_t Link_getBytesInPerMinute(T L, int count);
 
 
 /**
@@ -109,7 +109,7 @@ long long Link_getBytesInPerMinute(T L, int count);
  * @param L A Link object
  * @return Incoming bytes total or -1 if not available.
  */
-long long Link_getBytesInTotal(T L);
+int64_t Link_getBytesInTotal(T L);
 
 
 /**
@@ -127,7 +127,7 @@ double Link_getSaturationInPerSecond(T L);
  * range given by 'now - count' (count max = 24h)
  * @return Incoming bytes per hour or -1 if not available.
  */
-long long Link_getBytesInPerHour(T L, int count);
+int64_t Link_getBytesInPerHour(T L, int count);
 
 
 /**
@@ -135,7 +135,7 @@ long long Link_getBytesInPerHour(T L, int count);
  * @param L A Link object
  * @return Incoming packets per second or -1 if not available.
  */
-long long Link_getPacketsInPerSecond(T L);
+int64_t Link_getPacketsInPerSecond(T L);
 
 
 /**
@@ -145,7 +145,7 @@ long long Link_getPacketsInPerSecond(T L);
  * the range given by 'now - count' (count max = 60m)
  * @return Incoming packets per minute statistics or -1 if not available.
  */
-long long Link_getPacketsInPerMinute(T L, int count);
+int64_t Link_getPacketsInPerMinute(T L, int count);
 
 
 /**
@@ -155,7 +155,7 @@ long long Link_getPacketsInPerMinute(T L, int count);
  * the range given by 'now - count' (count max = 24h)
  * @return Incoming packets per hour statistics or -1 if not available.
  */
-long long Link_getPacketsInPerHour(T L, int count);
+int64_t Link_getPacketsInPerHour(T L, int count);
 
 
 /**
@@ -163,7 +163,7 @@ long long Link_getPacketsInPerHour(T L, int count);
  * @param L A Link object
  * @return Incoming packets total or -1 if not available.
  */
-long long Link_getPacketsInTotal(T L);
+int64_t Link_getPacketsInTotal(T L);
 
 
 /**
@@ -171,7 +171,7 @@ long long Link_getPacketsInTotal(T L);
  * @param L A Link object
  * @return Incoming errors per second or -1 if not available.
  */
-long long Link_getErrorsInPerSecond(T L);
+int64_t Link_getErrorsInPerSecond(T L);
 
 
 /**
@@ -181,7 +181,7 @@ long long Link_getErrorsInPerSecond(T L);
  * the range given by 'now - count' (count max = 60m)
  * @return Incoming errors per minute or -1 if not available.
  */
-long long Link_getErrorsInPerMinute(T L, int count);
+int64_t Link_getErrorsInPerMinute(T L, int count);
 
 
 /**
@@ -191,7 +191,7 @@ long long Link_getErrorsInPerMinute(T L, int count);
  * the range given by 'now - count' (count max = 24h)
  * @return Incoming errors per hour or -1 if not available.
  */
-long long Link_getErrorsInPerHour(T L, int count);
+int64_t Link_getErrorsInPerHour(T L, int count);
 
 
 /**
@@ -199,7 +199,7 @@ long long Link_getErrorsInPerHour(T L, int count);
  * @param L A Link object
  * @return Incoming errors total or -1 if not available.
  */
-long long Link_getErrorsInTotal(T L);
+int64_t Link_getErrorsInTotal(T L);
 
 
 /**
@@ -207,7 +207,7 @@ long long Link_getErrorsInTotal(T L);
  * @param L A Link object
  * @return Outgoing bytes per second or -1 if not available.
  */
-long long Link_getBytesOutPerSecond(T L);
+int64_t Link_getBytesOutPerSecond(T L);
 
 
 /**
@@ -217,7 +217,7 @@ long long Link_getBytesOutPerSecond(T L);
  * the range given by 'now - count' (count max = 60m)
  * @return Outgoing bytes per minute or -1 if not available.
  */
-long long Link_getBytesOutPerMinute(T L, int count);
+int64_t Link_getBytesOutPerMinute(T L, int count);
 
 
 /**
@@ -227,7 +227,7 @@ long long Link_getBytesOutPerMinute(T L, int count);
  * the range given by 'now - count' (count max = 24h)
  * @return Outgoing bytes per hour or -1 if not available.
  */
-long long Link_getBytesOutPerHour(T L, int count);
+int64_t Link_getBytesOutPerHour(T L, int count);
 
 
 /**
@@ -235,7 +235,7 @@ long long Link_getBytesOutPerHour(T L, int count);
  * @param L A Link object
  * @return Outgoing bytes total or -1 if not available.
  */
-long long Link_getBytesOutTotal(T L);
+int64_t Link_getBytesOutTotal(T L);
 
 
 /**
@@ -251,7 +251,7 @@ double Link_getSaturationOutPerSecond(T L);
  * @param L A Link object
  * @return Outgoing packets per second or -1 if not available.
  */
-long long Link_getPacketsOutPerSecond(T L);
+int64_t Link_getPacketsOutPerSecond(T L);
 
 
 /**
@@ -261,7 +261,7 @@ long long Link_getPacketsOutPerSecond(T L);
  * the range given by 'now - count' (count max = 60m)
  * @return Outgoing packets per minute or -1 if not available.
  */
-long long Link_getPacketsOutPerMinute(T L, int count);
+int64_t Link_getPacketsOutPerMinute(T L, int count);
 
 
 /**
@@ -271,7 +271,7 @@ long long Link_getPacketsOutPerMinute(T L, int count);
  * the range given by 'now - count' (count max = 24h)
  * @return Outgoing packets per hour or -1 if not available.
  */
-long long Link_getPacketsOutPerHour(T L, int count);
+int64_t Link_getPacketsOutPerHour(T L, int count);
 
 
 /**
@@ -279,7 +279,7 @@ long long Link_getPacketsOutPerHour(T L, int count);
  * @param L A Link object
  * @return Outgoing packets total or -1 if not available.
  */
-long long Link_getPacketsOutTotal(T L);
+int64_t Link_getPacketsOutTotal(T L);
 
 
 /**
@@ -287,7 +287,7 @@ long long Link_getPacketsOutTotal(T L);
  * @param L A Link object
  * @return Outgoing errors per second or -1 if not available.
  */
-long long Link_getErrorsOutPerSecond(T L);
+int64_t Link_getErrorsOutPerSecond(T L);
 
 
 /**
@@ -297,7 +297,7 @@ long long Link_getErrorsOutPerSecond(T L);
  * the range given by 'now - count' (count max = 60m)
  * @return Outgoing errors per minute or -1 if not available.
  */
-long long Link_getErrorsOutPerMinute(T L, int count);
+int64_t Link_getErrorsOutPerMinute(T L, int count);
 
 
 /**
@@ -307,7 +307,7 @@ long long Link_getErrorsOutPerMinute(T L, int count);
  * the range given by 'now - count' (count max = 24h)
  * @return Outgoing errors per hour or -1 if not available.
  */
-long long Link_getErrorsOutPerHour(T L, int count);
+int64_t Link_getErrorsOutPerHour(T L, int count);
 
 
 /**
@@ -315,7 +315,7 @@ long long Link_getErrorsOutPerHour(T L, int count);
  * @param L A Link object
  * @return Outgoing errors total or -1 if not available
  */
-long long Link_getErrorsOutTotal(T L);
+int64_t Link_getErrorsOutTotal(T L);
 
 
 /**
@@ -331,7 +331,7 @@ int Link_getState(T L);
  * @param L A Link object
  * @return Interface speed [bps] (-1 = N/A)
  */
-long long Link_getSpeed(T L);
+int64_t Link_getSpeed(T L);
 
 
 /**
